requirements.txt
# CodeIndex Python SDK - 用于查询代码符号和摘要
caicai-codeindex>=0.2.0

# 类型注解支持（Python < 3.9 兼容）
typing-extensions>=4.0.0

# YAML 配置文件支持
PyYAML>=6.0

# 日志和调试
loguru>=0.7.0
